Running rkt on master node

+++ cluster/saltbase/salt/kube-addons/kube-addons.sh
@@ -25,13 +25,20 @@ SYSTEM_NAMESPACE=kube-system
 trusty_master=${TRUSTY_MASTER:-false}
 
 function ensure_python() {
-  if ! python --version > /dev/null 2>&1; then    
+  if ! python --version > /dev/null 2>&1; then 
     echo "No python on the machine, will use a python image"
-    local -r PYTHON_IMAGE=python:2.7-slim-pyyaml
-    export PYTHON="docker run --interactive --rm --net=none ${PYTHON_IMAGE} python"
+    if [[ "${MASTER_CONTAINER_RUNTIME}" == "rkt" ]]; then
+      local -r PYTHON_IMAGE=$(/opt/rkt/rkt image list | grep python:2.7-slim-pyyaml | cut -f 1)
+      export PYTHON="/opt/rkt/rkt run --interactive --net=none --stage1-path=/opt/rkt/stage1-fly.aci --insecure-options=image ${PYTHON_IMAGE} --exec=/usr/local/bin/python -- "
+    else
+      local -r PYTHON_IMAGE=python:2.7-slim-pyyaml
+      export PYTHON="docker run --interactive --rm --net=none ${PYTHON_IMAGE} python"
+    fi
   else
     export PYTHON=python
   fi
+
+  echo "PYTHON=$PYTHON"
 }
 
 # $1 filename of addon to start.
@@ -91,6 +98,17 @@ function load-docker-images() {
   done
 }
 
+function load-rkt-images() {
+	for image in "$1/"*; do
+		(cd /tmp ; /opt/rkt/docker2aci "${image}")
+	done
+	for aci in "/tmp/"*.aci; do
+		/opt/rkt/rkt fetch "${aci}" --insecure-options=image
+	done
+
+	/opt/rkt/rkt fetch coreos.com/rkt/stage1-fly:1.3.0
+}
+
 # The business logic for whether a given object should be created
 # was already enforced by salt, and /etc/kubernetes/addons is the
 # managed result is of that. Start everything below that directory.
@@ -98,8 +116,15 @@ echo "== Kubernetes addon manager started at $(date -Is) with ADDON_CHECK_INTERV
 
 # Load any images that we may need. This is not needed for trusty master and
 # the way it restarts docker daemon does not work for trusty.
-if [[ "${trusty_master}" == "false" ]]; then
-  load-docker-images /srv/salt/kube-addons-images
+if [[ "${MASTER_CONTAINER_RUNTIME}" == "rkt" ]]; then
+  echo "rkt"
+  load-rkt-images /srv/salt/kube-addons-images
+else
+  echo "not rkt"
+  if [[ "${trusty_master}" == "false" ]]; then
+    echo "docker"
+    load-docker-images /srv/salt/kube-addons-images
+  fi
 fi
 
 ensure_python
